refactor: move template page handler
